@@ -10,7 +10,10 @@
     "babel-loader": "^5.3.2",
     "classnames": "^2.1.3",
     "css-loader": "^0.16.0",
-    "react": "^0.14.0",
+    "prop-types": "^15.5.8",
+    "react": "15.6.1",
+    "react-button": "^1.2.1",
+    "react-dom": "^15.6.1",
     "react-treeview": "^0.4.0",
     "simple-ajax": "^2.1.0",
     "style-loader": "^0.12.3",
