@@ -7,7 +7,11 @@
  * of patent rights can be found in the PATENTS file in the same directory.
  */
 
+import PropTypes from 'prop-types';
 import React from 'react';
+
+import HTTP from 'js/http';
+var Button = require('react-button');
 
 require('css/inspector.css');
 
@@ -37,6 +41,11 @@ class Inspector extends React.Component {
     }
 
     const attributes = this.props.selectedNode.attributes;
+    const tapButton =
+      <Button onClick={(event) => this.tap(this.props.selectedNode)}>
+        tap
+      </Button>;
+
     return (
       <div>
         {this.renderField('Class', attributes.type)}
@@ -47,13 +56,20 @@ class Inspector extends React.Component {
         {this.renderField('Rect', attributes.rect)}
         {this.renderField('isEnabled', boolToString(attributes.isEnabled))}
         {this.renderField('isVisible', boolToString(attributes.isVisible))}
-      </div>
+        {this.renderField('Tap', tapButton, false)}
+     </div>
     );
   }
 
-  renderField(fieldName, fieldValue) {
+  renderField(fieldName, fieldValue, castToString = true) {
     if (fieldValue == null) {
       return null;
+    }
+    var value;
+    if (castToString) {
+      value = String(fieldValue);
+    } else {
+      value = fieldValue;
     }
     return (
       <div className="inspector-field">
@@ -61,15 +77,42 @@ class Inspector extends React.Component {
           {fieldName}:
         </div>
         <div className="inspector-field-value">
-          {String(fieldValue)}
+          {value}
         </div>
       </div>
+    );
+  }
+
+  tap(node) {
+    HTTP.get(
+      'status', (status_result) => {
+        var session_id = status_result.sessionId;
+        HTTP.post(
+          'session/' + session_id + '/elements',
+          JSON.stringify({
+            'using': 'link text',
+            'value': 'label=' + node.attributes.label,
+          }),
+          (elements_result) => {
+            var elements = elements_result.value;
+            var element_id = elements[0].ELEMENT;
+
+            HTTP.post(
+              'session/' + session_id + '/element/' + element_id + '/click',
+              JSON.stringify({}),
+              (result) => {
+                this.props.refreshApp();
+              },
+            );
+          },
+        );
+      },
     );
   }
 }
 
 Inspector.propTypes = {
-  selectedNode: React.PropTypes.object,
+  selectedNode: PropTypes.object,
 };
 
 module.exports = Inspector;
