@@ -7,7 +7,13 @@
  * of patent rights can be found in the PATENTS file in the same directory.
  */
 
+import PropTypes from 'prop-types';
 import React from 'react';
+
+import HTTP from 'js/http';
+import GestureRecognizer from 'js/gesture_recognizer';
+
+var Button = require('react-button');
 
 require('css/screen.css');
 
@@ -18,6 +24,14 @@ class Screen extends React.Component {
         <div className="section-caption">
           Screen
         </div>
+        <div>
+          <Button onClick={(ev) => {this.home(ev); }} >
+            Home
+          </Button>
+          <Button onClick={this.props.refreshApp} >
+            Refresh
+          </Button>
+        </div>
         <div className="section-content-container">
           <div className="screen-screenshot-container"
             style={this.styleWithScreenSize()}>
@@ -27,6 +41,20 @@ class Screen extends React.Component {
         </div>
       </div>
     );
+  }
+
+  gestureRecognizer() {
+    if (!this._gestureRecognizer) {
+      this._gestureRecognizer = new GestureRecognizer({
+        onClick: (ev) => {
+          this.onScreenShotClick(ev);
+        },
+        onDrag: (params) => {
+          this.onScreenShotDrag(params);
+        },
+      });
+    }
+    return this._gestureRecognizer;
   }
 
   styleWithScreenSize() {
@@ -41,14 +69,103 @@ class Screen extends React.Component {
     return this.props.screenshot ? this.props.screenshot : {};
   }
 
+  onScreenShotDrag(params) {
+    var fromX = params.origin.x - document.getElementById('screenshot').offsetLeft;
+    var fromY = params.origin.y - document.getElementById('screenshot').offsetTop;
+    var toX = params.end.x - document.getElementById('screenshot').offsetLeft;
+    var toY = params.end.y - document.getElementById('screenshot').offsetTop;
+
+    fromX = this.scaleCoord(fromX);
+    fromY = this.scaleCoord(fromY);
+    toX = this.scaleCoord(toX);
+    toY = this.scaleCoord(toY);
+
+    HTTP.get(
+      'status', (status_result) => {
+        var session_id = status_result.sessionId;
+        HTTP.post(
+          'session/' + session_id + '/wda/element/0/dragfromtoforduration',
+          JSON.stringify({
+            'fromX': fromX,
+            'fromY': fromY,
+            'toX': toX,
+            'toY': toY,
+            'duration': params.duration,
+          }),
+          (tap_result) => {
+            this.props.refreshApp();
+          },
+        );
+      },
+    );
+  }
+
+  scaleCoord(coord) {
+    var screenshot = this.screenshot();
+    var pxPtScale = screenshot.width / this.props.rootNode.rect.size.width;
+    return coord / screenshot.scale / pxPtScale;
+  }
+
+  onScreenShotClick(point) {
+    var x = point.x - document.getElementById('screenshot').offsetLeft;
+    var y = point.y - document.getElementById('screenshot').offsetTop;
+    x = this.scaleCoord(x);
+    y = this.scaleCoord(y);
+
+    HTTP.get(
+      'status', (status_result) => {
+        var session_id = status_result.sessionId;
+        HTTP.post(
+          'session/' + session_id + '/wda/tap/0',
+          JSON.stringify({
+            'x': x,
+            'y': y,
+          }),
+          (tap_result) => {
+            this.props.refreshApp();
+          },
+        );
+      },
+    );
+  }
+
+  onMouseDown(ev) {
+    this.gestureRecognizer().onMouseDown(ev);
+  }
+
+  onMouseMove(ev) {
+    this.gestureRecognizer().onMouseMove(ev);
+  }
+
+  onMouseUp(ev) {
+    this.gestureRecognizer().onMouseUp(ev);
+  }
+
+  home(ev) {
+    HTTP.post(
+      '/wda/homescreen',
+      JSON.stringify({}),
+      (result) => {
+        this.props.refreshApp();
+      },
+    );
+  }
+
   renderScreenshot() {
     return (
       <img
         className="screen-screenshot"
         src={this.screenshot().source}
-        style={this.styleWithScreenSize()} />
+        style={this.styleWithScreenSize()}
+        onMouseDown={(ev) => this.onMouseDown(ev)}
+        onMouseMove={(ev) => this.onMouseMove(ev)}
+        onMouseUp={(ev) => this.onMouseUp(ev)}
+        draggable="false"
+        id="screenshot"
+      />
     );
   }
+
 
   renderHighlightedNode() {
     if (this.props.highlightedNode == null) {
@@ -90,8 +207,8 @@ class Screen extends React.Component {
 }
 
 Screen.propTypes = {
-  highlightedNode: React.PropTypes.object,
-  screenshot: React.PropTypes.object,
+  highlightedNode: PropTypes.object,
+  screenshot: PropTypes.object,
 };
 
 module.exports = Screen;
