@@ -7,6 +7,7 @@
  * of patent rights can be found in the PATENTS file in the same directory.
  */
 
+import PropTypes from 'prop-types';
 import React from 'react';
 import TreeView from 'react-treeview';
 
@@ -117,10 +118,10 @@ class Tree extends React.Component {
 }
 
 Tree.propTypes = {
-  onSelectedNodeChange: React.PropTypes.func,
-  onHighlightedNodeChange: React.PropTypes.func,
-  rootNode: React.PropTypes.object,
-  selectedNode: React.PropTypes.object,
+  onSelectedNodeChange: PropTypes.func,
+  onHighlightedNodeChange: PropTypes.func,
+  rootNode: PropTypes.object,
+  selectedNode: PropTypes.object,
 };
 
 module.exports = Tree;
