@@ -4,22 +4,22 @@
 //     class-dump is Copyright (C) 1997-1998, 2000-2001, 2004-2013 by Steve Nygard.
 //
 
-@class NSArray, NSMutableArray, NSString;
+@class NSArray, NSMutableArray, NSString, XCPointerEventPath;
 
 @interface XCSynthesizedEventRecord : NSObject <NSSecureCoding>
 {
     NSMutableArray *_eventPaths;
     NSString *_name;
-    long long _interfaceOrientation;
+    UIInterfaceOrientation _interfaceOrientation;
 }
 @property(readonly) UIInterfaceOrientation interfaceOrientation; // @synthesize interfaceOrientation=_interfaceOrientation;
 @property(readonly, copy) NSString *name; // @synthesize name=_name;
 @property(readonly) double maximumOffset;
 @property(readonly) NSArray *eventPaths;
 
-- (void)addPointerEventPath:(id)arg1;
-- (id)initWithName:(id)arg1 interfaceOrientation:(long long)arg2;
+- (void)addPointerEventPath:(XCPointerEventPath *)arg1;
+- (id)initWithName:(NSString *)arg1 interfaceOrientation:(UIInterfaceOrientation)arg2;
 - (id)init;
-- (BOOL)synthesizeWithError:(id *)arg1;
+- (BOOL)synthesizeWithError:(NSError **)arg1;
 
 @end
