@@ -15,6 +15,7 @@
 #import "FBLogger.h"
 #import "FBMacros.h"
 #import "FBMathUtils.h"
+#import "FBPredicate.h"
 #import "XCElementSnapshot+FBHelpers.h"
 #import "XCElementSnapshot.h"
 #import "XCEventGenerator.h"
@@ -25,10 +26,10 @@
 #import "XCUICoordinate.h"
 #import "XCUIElement+FBIsVisible.h"
 #import "XCUIElement.h"
+#import "XCUIElement+FBUtilities.h"
 #import "XCUIElement+FBWebDriverAttributes.h"
 
 const CGFloat FBFuzzyPointThreshold = 20.f; //Smallest determined value that is not interpreted as touch
-const CGFloat FBFullscreenNormalizedDistance = 1.0f;
 const CGFloat FBScrollToVisibleNormalizedDistance = .5f;
 const CGFloat FBScrollVelocity = 200.f;
 const CGFloat FBScrollBoundingVelocityPadding = 0.0f;
@@ -49,24 +50,24 @@ const CGFloat FBMinimumTouchEventDelay = 0.1f;
 
 @implementation XCUIElement (FBScrolling)
 
-- (void)fb_scrollUp
+- (void)fb_scrollUpByNormalizedDistance:(CGFloat)distance
 {
-  [self.lastSnapshot fb_scrollUpByNormalizedDistance:FBFullscreenNormalizedDistance];
+  [self.fb_lastSnapshot fb_scrollUpByNormalizedDistance:distance];
 }
 
-- (void)fb_scrollDown
+- (void)fb_scrollDownByNormalizedDistance:(CGFloat)distance
 {
-  [self.lastSnapshot fb_scrollDownByNormalizedDistance:FBFullscreenNormalizedDistance];
+  [self.fb_lastSnapshot fb_scrollDownByNormalizedDistance:distance];
 }
 
-- (void)fb_scrollLeft
+- (void)fb_scrollLeftByNormalizedDistance:(CGFloat)distance
 {
-  [self.lastSnapshot fb_scrollLeftByNormalizedDistance:FBFullscreenNormalizedDistance];
+  [self.fb_lastSnapshot fb_scrollLeftByNormalizedDistance:distance];
 }
 
-- (void)fb_scrollRight
+- (void)fb_scrollRightByNormalizedDistance:(CGFloat)distance
 {
-  [self.lastSnapshot fb_scrollRightByNormalizedDistance:FBFullscreenNormalizedDistance];
+  [self.fb_lastSnapshot fb_scrollRightByNormalizedDistance:distance];
 }
 
 - (BOOL)fb_scrollToVisibleWithError:(NSError **)error
@@ -95,8 +96,8 @@ const CGFloat FBMinimumTouchEventDelay = 0.1f;
                                @(XCUIElementTypeTable),
                                @(XCUIElementTypeWebView),
                                ];
-
-  XCElementSnapshot *scrollView = [self.lastSnapshot fb_parentMatchingOneOfTypes:acceptedParents
+  XCElementSnapshot *elementSnapshot = self.fb_lastSnapshot;
+  XCElementSnapshot *scrollView = [elementSnapshot fb_parentMatchingOneOfTypes:acceptedParents
       filter:^(XCElementSnapshot *snapshot) {
 
          if (![snapshot isWDVisible]) {
@@ -105,7 +106,7 @@ const CGFloat FBMinimumTouchEventDelay = 0.1f;
 
          cellSnapshots = [snapshot fb_descendantsCellSnapshots];
 
-         visibleCellSnapshots = [cellSnapshots filteredArrayUsingPredicate:[NSPredicate predicateWithFormat:@"%K == YES", FBStringify(XCUIElement, fb_isVisible)]];
+         visibleCellSnapshots = [cellSnapshots filteredArrayUsingPredicate:[FBPredicate predicateWithFormat:@"%K == YES", FBStringify(XCUIElement, fb_isVisible)]];
 
          if (visibleCellSnapshots.count > 1) {
            return YES;
@@ -120,7 +121,7 @@ const CGFloat FBMinimumTouchEventDelay = 0.1f;
      buildError:error];
   }
 
-  XCElementSnapshot *targetCellSnapshot = [self.lastSnapshot fb_parentCellSnapshot];
+  XCElementSnapshot *targetCellSnapshot = [elementSnapshot fb_parentCellSnapshot];
 
   XCElementSnapshot *lastSnapshot = visibleCellSnapshots.lastObject;
   // Can't just do indexOfObject, because targetCellSnapshot may represent the same object represented by a member of cellSnapshots, yet be a different object
@@ -148,7 +149,7 @@ const CGFloat FBMinimumTouchEventDelay = 0.1f;
   const NSUInteger maxScrollCount = 25;
   NSUInteger scrollCount = 0;
 
-  XCElementSnapshot *prescrollSnapshot = self.lastSnapshot;
+  XCElementSnapshot *prescrollSnapshot = self.fb_lastSnapshot;
   // Scrolling till cell is visible and get current value of frames
   while (![self fb_isEquivalentElementSnapshotVisible:prescrollSnapshot] && scrollCount < maxScrollCount) {
     if (targetCellIndex < visibleCellIndex) {
@@ -171,7 +172,7 @@ const CGFloat FBMinimumTouchEventDelay = 0.1f;
   }
 
   // Cell is now visible, but it might be only partialy visible, scrolling till whole frame is visible
-  targetCellSnapshot = [self.lastSnapshot fb_parentCellSnapshot];
+  targetCellSnapshot = [self.fb_lastSnapshot fb_parentCellSnapshot];
   CGVector scrollVector = CGVectorMake(targetCellSnapshot.visibleFrame.size.width - targetCellSnapshot.frame.size.width,
                                        targetCellSnapshot.visibleFrame.size.height - targetCellSnapshot.frame.size.height
                                        );
@@ -186,8 +187,7 @@ const CGFloat FBMinimumTouchEventDelay = 0.1f;
   if (self.fb_isVisible) {
     return YES;
   }
-  [self.application resolve];
-  for (XCElementSnapshot *elementSnapshot in self.application.lastSnapshot._allDescendants.copy) {
+  for (XCElementSnapshot *elementSnapshot in self.application.fb_lastSnapshot._allDescendants.copy) {
     // We are comparing pre-scroll snapshot so frames are irrelevant.
     if ([snapshot fb_framelessFuzzyMatchesElement:elementSnapshot] && elementSnapshot.fb_isVisible) {
       return YES;
@@ -273,7 +273,7 @@ const CGFloat FBMinimumTouchEventDelay = 0.1f;
     return YES;
   }
 
-  double offset = 0.3; // Waiting before scrolling helps to make it more stable
+  CGFloat offset = 0.3f; // Waiting before scrolling helps to make it more stable
   double scrollingTime = MAX(fabs(vector.dx), fabs(vector.dy))/FBScrollVelocity;
   XCPointerEventPath *touchPath = [[XCPointerEventPath alloc] initForTouchAtPoint:startCoordinate.screenPoint offset:offset];
   offset += MAX(scrollingTime, FBMinimumTouchEventDelay); // Setting Minimum scrolling time to avoid testmanager complaining about timing
@@ -286,7 +286,7 @@ const CGFloat FBMinimumTouchEventDelay = 0.1f;
 
   __block BOOL didSucceed = NO;
   __block NSError *innerError;
-  [FBRunLoopSpinner spinUntilCompletion:^(void(^completion)()){
+  [FBRunLoopSpinner spinUntilCompletion:^(void(^completion)(void)){
     [[FBXCTestDaemonsProxy testRunnerProxy] _XCT_synthesizeEvent:event completion:^(NSError *scrollingError) {
       innerError = scrollingError;
       didSucceed = (scrollingError == nil);
