@@ -10,11 +10,17 @@
 #import "XCUIElement+FBUtilities.h"
 
 #import "FBAlert.h"
+#import "FBLogger.h"
+#import "FBMacros.h"
 #import "FBMathUtils.h"
+#import "FBPredicate.h"
 #import "FBRunLoopSpinner.h"
+#import "XCAXClient_iOS.h"
 #import "XCUIElement+FBWebDriverAttributes.h"
 
 @implementation XCUIElement (FBUtilities)
+
+static const NSTimeInterval FBANIMATION_TIMEOUT = 5.0;
 
 - (BOOL)fb_waitUntilFrameIsStable
 {
@@ -42,12 +48,12 @@
   if (!self.exists) {
     return NO;
   }
-  [self resolve];
-  [element resolve];
-  if ([self.lastSnapshot _isAncestorOfElement:element.lastSnapshot]) {
+  XCElementSnapshot *snapshot = self.fb_lastSnapshot;
+  XCElementSnapshot *elementSnapshot = element.fb_lastSnapshot;
+  if ([snapshot _isAncestorOfElement:elementSnapshot]) {
     return NO;
   }
-  if ([self.lastSnapshot _matchesElement:element.lastSnapshot]) {
+  if ([snapshot _matchesElement:elementSnapshot]) {
     return NO;
   }
   return YES;
@@ -55,22 +61,61 @@
 
 - (XCElementSnapshot *)fb_lastSnapshot
 {
-  if (self.lastSnapshot) {
-    return self.lastSnapshot;
-  }
+  [self query];
   [self resolve];
   return self.lastSnapshot;
 }
 
-- (NSDictionary<NSNumber *, NSArray<XCUIElement *> *> *)fb_categorizeDescendants:(NSSet<NSNumber *> *)byTypes
+- (NSArray<XCUIElement *> *)fb_filterDescendantsWithSnapshots:(NSArray<XCElementSnapshot *> *)snapshots
 {
-  NSMutableDictionary *result = [NSMutableDictionary dictionary];
-  [byTypes enumerateObjectsUsingBlock:^(NSNumber *elementTypeAsNumber, BOOL *stopEnum) {
-    XCUIElementType elementType = (XCUIElementType)elementTypeAsNumber.unsignedIntegerValue;
-    NSArray *descendantsOfType = [[self descendantsMatchingType:elementType] allElementsBoundByIndex];
-    result[elementTypeAsNumber] = descendantsOfType;
+  if (0 == snapshots.count) {
+    return @[];
+  }
+  NSArray<NSNumber *> *matchedUids = [snapshots valueForKey:FBStringify(XCUIElement, wdUID)];
+  NSMutableArray<XCUIElement *> *matchedElements = [NSMutableArray array];
+  if ([matchedUids containsObject:@(self.wdUID)]) {
+    if (1 == snapshots.count) {
+      return @[self];
+    }
+    [matchedElements addObject:self];
+  }
+  XCUIElementType type = XCUIElementTypeAny;
+  NSArray<NSNumber *> *uniqueTypes = [snapshots valueForKeyPath:[NSString stringWithFormat:@"@distinctUnionOfObjects.%@", FBStringify(XCUIElement, elementType)]];
+  if (uniqueTypes && [uniqueTypes count] == 1) {
+    type = [uniqueTypes.firstObject intValue];
+  }
+  [matchedElements addObjectsFromArray:[[self descendantsMatchingType:type] matchingPredicate:[FBPredicate predicateWithFormat:@"%K IN %@", FBStringify(XCUIElement, wdUID), matchedUids]].allElementsBoundByIndex];
+  if (matchedElements.count <= 1) {
+    // There is no need to sort elements if count of matches is not greater than one
+    return matchedElements;
+  }
+  NSMutableArray<XCUIElement *> *sortedElements = [NSMutableArray array];
+  [snapshots enumerateObjectsUsingBlock:^(XCElementSnapshot *snapshot, NSUInteger snapshotIdx, BOOL *stopSnapshotEnum) {
+    XCUIElement *matchedElement = nil;
+    for (XCUIElement *element in matchedElements) {
+      if (element.wdUID == snapshot.wdUID) {
+        matchedElement = element;
+        break;
+      }
+    }
+    if (matchedElement) {
+      [sortedElements addObject:matchedElement];
+      [matchedElements removeObject:matchedElement];
+    }
   }];
-  return result.copy;
+  return sortedElements.copy;
+}
+
+- (BOOL)fb_waitUntilSnapshotIsStable
+{
+  dispatch_semaphore_t sem = dispatch_semaphore_create(0);
+  [[XCAXClient_iOS sharedClient] notifyWhenNoAnimationsAreActiveForApplication:self.application reply:^{dispatch_semaphore_signal(sem);}];
+  dispatch_time_t timeout = dispatch_time(DISPATCH_TIME_NOW, (int64_t)(FBANIMATION_TIMEOUT * NSEC_PER_SEC));
+  BOOL result = 0 == dispatch_semaphore_wait(sem, timeout);
+  if (!result) {
+    [FBLogger logFmt:@"There are still some active animations in progress after %.2f seconds timeout. Visibility detection may cause unexpected delays.", FBANIMATION_TIMEOUT];
+  }
+  return result;
 }
 
 @end
