@@ -12,6 +12,7 @@
 #import "FBApplicationProcessProxy.h"
 #import "FBRunLoopSpinner.h"
 #import "FBMacros.h"
+#import "FBXCodeCompatibility.h"
 #import "XCAccessibilityElement.h"
 #import "XCAXClient_iOS.h"
 #import "XCUIApplication.h"
@@ -38,7 +39,7 @@
   if (!activeApplicationElement) {
     return nil;
   }
-  FBApplication *application = [FBApplication appWithPID:activeApplicationElement.processIdentifier];
+  FBApplication *application = [FBApplication fb_applicationWithPID:activeApplicationElement.processIdentifier];
   NSAssert(nil != application, @"Active application instance is not expected to be equal to nil", nil);
   [application query];
   [application resolve];
@@ -55,6 +56,20 @@
     return application;
   }
   application = [super appWithPID:processID];
+  [FBApplication fb_registerApplication:application withProcessID:processID];
+  return application;
+}
+
++ (instancetype)applicationWithPID:(pid_t)processID
+{
+  if ([NSProcessInfo processInfo].processIdentifier == processID) {
+    return nil;
+  }
+  FBApplication *application = [self fb_registeredApplicationWithProcessID:processID];
+  if (application) {
+    return application;
+  }
+  application = [super applicationWithPID:processID];
   [FBApplication fb_registerApplication:application withProcessID:processID];
   return application;
 }
