@@ -33,7 +33,10 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, readonly, copy) NSString *wdType;
 
 /*! Element's value */
-@property (nonatomic, readonly, strong, nullable) id wdValue;
+@property (nonatomic, readonly, strong, nullable) NSString *wdValue;
+
+/*! Element's unique identifier */
+@property (nonatomic, readonly) NSUInteger wdUID;
 
 /*! Whether element is enabled */
 @property (nonatomic, readonly, getter = isWDEnabled) BOOL wdEnabled;
