@@ -48,7 +48,8 @@ static NSString *const OBJC_PROP_ATTRIBS_SEPARATOR = @",";
   dispatch_once(&onceToken, ^{
     NSMutableDictionary *wdPropertyGettersMapping = [NSMutableDictionary new];
     unsigned int propsCount = 0;
-    objc_property_t *properties = protocol_copyPropertyList(objc_getProtocol("FBElement"), &propsCount);
+    Protocol * aProtocol = objc_getProtocol(protocol_getName(@protocol(FBElement)));
+    objc_property_t *properties = protocol_copyPropertyList(aProtocol, &propsCount);
     for (unsigned int i = 0; i < propsCount; ++i) {
       objc_property_t property = properties[i];
       const char *name = property_getName(property);
@@ -87,9 +88,13 @@ static NSString *const OBJC_PROP_ATTRIBS_SEPARATOR = @",";
         aliasName = [NSString stringWithFormat:@"%@",
                         [propName substringWithRange:NSMakeRange(WD_PREFIX.length, 1)].lowercaseString];
       } else {
-        aliasName = [NSString stringWithFormat:@"%@%@",
-                        [propName substringWithRange:NSMakeRange(WD_PREFIX.length, 1)].lowercaseString,
-                        [propName substringFromIndex:WD_PREFIX.length + 1]];
+        NSString *propNameWithoutPrefix = [propName substringFromIndex:WD_PREFIX.length];
+        NSString *firstPropNameCharacter = [propNameWithoutPrefix substringWithRange:NSMakeRange(0, 1)];
+        if (![propNameWithoutPrefix isEqualToString:[propNameWithoutPrefix uppercaseString]]) {
+          // Lowercase the first character for the alias if the property name is not an uppercase abbreviation
+          firstPropNameCharacter = firstPropNameCharacter.lowercaseString;
+        }
+        aliasName = [NSString stringWithFormat:@"%@%@", firstPropNameCharacter, [propNameWithoutPrefix substringFromIndex:1]];
       }
       if ([[wdPropertyGettersMapping valueForKey:propName] isKindOfClass:NSNull.class]) {
         // no getter
