@@ -10,18 +10,43 @@
 #import <Foundation/Foundation.h>
 
 @class RouteResponse, RoutingHTTPServer, FBExceptionHandler;
+@protocol FBWebServerDelegate;
 
 NS_ASSUME_NONNULL_BEGIN
 
 /**
- HTTP and USB service wrapper, that handel's requests and responses
+ HTTP and USB service wrapper, handling requests and responses
  */
 @interface FBWebServer : NSObject
+
+/**
+ Server delegate.
+ */
+@property (weak, nonatomic) id<FBWebServerDelegate> delegate;
 
 /**
  Starts WebDriverAgent service by booting HTTP and USB server
  */
 - (void)startServing;
+
+/**
+ Stops WebDriverAgent service, shutting down HTTP and USB servers.
+ */
+- (void)stopServing;
+
+@end
+
+/**
+ The protocol allowing the server delegate to handle messages from the server.
+ */
+@protocol FBWebServerDelegate <NSObject>
+
+/**
+ The server requested WebDriverAgent service shutdown.
+
+ @param webServer Server instance.
+ */
+- (void)webServerDidRequestShutdown:(FBWebServer *)webServer;
 
 @end
 
