@@ -14,21 +14,20 @@
 #include "TargetConditionals.h"
 #import "XCTestPrivateSymbols.h"
 
-BOOL _AXSAutomationSetFauxCollectionViewCellsEnabled(BOOL);
-
 static NSUInteger const DefaultStartingPort = 8100;
 static NSUInteger const DefaultPortRange = 100;
 
 static BOOL FBShouldUseTestManagerForVisibilityDetection = NO;
+static BOOL FBShouldUseCompactResponses = YES;
 static NSUInteger FBMaxTypingFrequency = 60;
 
 @implementation FBConfiguration
 
 #pragma mark Public
 
-+ (void)shouldShowFakeCollectionViewCells:(BOOL)showFakeCells
++ (void)disableRemoteQueryEvaluation
 {
-  _AXSAutomationSetFauxCollectionViewCellsEnabled(showFakeCells);
+  [[NSUserDefaults standardUserDefaults] setBool:YES forKey:@"XCTDisableRemoteQueryEvaluation"];
 }
 
 + (NSRange)bindingPortRange
@@ -59,6 +58,16 @@ static NSUInteger FBMaxTypingFrequency = 60;
 + (BOOL)shouldUseTestManagerForVisibilityDetection
 {
   return FBShouldUseTestManagerForVisibilityDetection;
+}
+
++ (void)setShouldUseCompactResponses:(BOOL)value
+{
+  FBShouldUseCompactResponses = value;
+}
+
++ (BOOL)shouldUseCompactResponses
+{
+  return FBShouldUseCompactResponses;
 }
 
 + (void)setMaxTypingFrequency:(NSUInteger)value
