@@ -42,14 +42,14 @@ FOUNDATION_EXPORT const unsigned char WebDriverAgentLib_VersionString[];
 #import <WebDriverAgentLib/FBSpringboardApplication.h>
 #import <WebDriverAgentLib/FBXPathCreator.h>
 #import <WebDriverAgentLib/FBWebServer.h>
-#import <WebDriverAgentLib/XCElementSnapshot+Helpers.h>
+#import <WebDriverAgentLib/XCElementSnapshot+FBHelpers.h>
 #import <WebDriverAgentLib/XCUIApplication+FBHelpers.h>
 #import <WebDriverAgentLib/XCUIDevice+FBHelpers.h>
-#import <WebDriverAgentLib/XCUIDevice+Rotation.h>
+#import <WebDriverAgentLib/XCUIDevice+FBRotation.h>
 #import <WebDriverAgentLib/XCUIElement+FBAccessibility.h>
 #import <WebDriverAgentLib/XCUIElement+FBFind.h>
 #import <WebDriverAgentLib/XCUIElement+FBIsVisible.h>
 #import <WebDriverAgentLib/XCUIElement+FBScrolling.h>
 #import <WebDriverAgentLib/XCUIElement+FBTap.h>
-#import <WebDriverAgentLib/XCUIElement+Utilities.h>
-#import <WebDriverAgentLib/XCUIElement+WebDriverAttributes.h>
+#import <WebDriverAgentLib/XCUIElement+FBUtilities.h>
+#import <WebDriverAgentLib/XCUIElement+FBWebDriverAttributes.h>
