@@ -22,7 +22,17 @@
 - (void)setUp
 {
   [super setUp];
-  [self goToAlertsPage];
+  static dispatch_once_t onceToken;
+  dispatch_once(&onceToken, ^{
+    [self launchApplication];
+    [self goToAlertsPage];
+  });
+}
+
+- (void)tearDown
+{
+  [super tearDown];
+  [[FBAlert alertWithApplication:self.testedApplication] dismissWithError:nil];
 }
 
 - (void)showApplicationAlert
@@ -57,6 +67,25 @@
   [self showApplicationAlert];
   XCTAssertTrue([alert.text containsString:@"Magic"]);
   XCTAssertTrue([alert.text containsString:@"Should read"]);
+}
+
+- (void)testAlertLabels
+{
+  FBAlert* alert = [FBAlert alertWithApplication:self.testedApplication];
+  XCTAssertNil(alert.buttonLabels);
+  [self showApplicationAlert];
+  XCTAssertNotNil(alert.buttonLabels);
+  XCTAssertEqual(1, alert.buttonLabels.count);
+  XCTAssertEqualObjects(@"Will do", alert.buttonLabels[0]);
+}
+
+- (void)testClickAlertButton
+{
+  FBAlert* alert = [FBAlert alertWithApplication:self.testedApplication];
+  XCTAssertFalse([alert clickAlertButton:@"Invalid" error:nil]);
+  [self showApplicationAlert];
+  XCTAssertFalse([alert clickAlertButton:@"Invalid" error:nil]);
+  XCTAssertTrue([alert clickAlertButton:@"Will do" error:nil]);
 }
 
 - (void)testAcceptingAlert
