@@ -11,6 +11,7 @@
 
 #import "FBIntegrationTestCase.h"
 
+#import "FBAlert.h"
 #import "FBElementCache.h"
 #import "FBTestMacros.h"
 #import "XCUIDevice+FBRotation.h"
@@ -26,12 +27,27 @@
 
 - (void)verifyTapWithOrientation:(UIDeviceOrientation)orientation
 {
-  [self goToAlertsPage];
   [[XCUIDevice sharedDevice] fb_setDeviceInterfaceOrientation:orientation];
   NSError *error;
   XCTAssertTrue(self.testedApplication.alerts.count == 0);
   [self.testedApplication.buttons[FBShowAlertButtonName] fb_tapWithError:&error];
   FBAssertWaitTillBecomesTrue(self.testedApplication.alerts.count > 0);
+}
+
+- (void)setUp
+{
+  [super setUp];
+  static dispatch_once_t onceToken;
+  dispatch_once(&onceToken, ^{
+    [self launchApplication];
+    [self goToAlertsPage];
+  });
+}
+
+- (void)tearDown
+{
+  [super tearDown];
+  [[FBAlert alertWithApplication:self.testedApplication] dismissWithError:nil];
 }
 
 - (void)testTap
@@ -56,7 +72,6 @@
 
 - (void)verifyTapByCoordinatesWithOrientation:(UIDeviceOrientation)orientation
 {
-  [self goToAlertsPage];
   [[XCUIDevice sharedDevice] fb_setDeviceInterfaceOrientation:orientation];
   NSError *error;
   XCTAssertTrue(self.testedApplication.alerts.count == 0);
