@@ -13,8 +13,9 @@
 #import "FBTestMacros.h"
 #import "XCElementSnapshot+FBHelpers.h"
 #import "XCUIElement.h"
-#import "XCUIElement+FBWebDriverAttributes.h"
 #import "XCUIElement+FBIsVisible.h"
+#import "XCUIElement+FBUtilities.h"
+#import "XCUIElement+FBWebDriverAttributes.h"
 
 @interface XCElementSnapshotHelperTests : FBIntegrationTestCase
 @property (nonatomic, strong) XCUIElement *testedView;
@@ -25,6 +26,10 @@
 - (void)setUp
 {
   [super setUp];
+  static dispatch_once_t onceToken;
+  dispatch_once(&onceToken, ^{
+    [self launchApplication];
+  });
   self.testedView = self.testedApplication.otherElements[@"MainView"];
   XCTAssertTrue(self.testedView.exists);
   [self.testedView resolve];
@@ -38,7 +43,7 @@
     @"Scrolling",
     @"Deadlock app",
   ]];
-  NSArray<XCElementSnapshot *> *matchingSnapshots = [self.testedView.lastSnapshot fb_descendantsMatchingType:XCUIElementTypeButton];
+  NSArray<XCElementSnapshot *> *matchingSnapshots = [self.testedView.fb_lastSnapshot fb_descendantsMatchingType:XCUIElementTypeButton];
   XCTAssertEqual(matchingSnapshots.count, expectedLabels.count);
   NSArray<NSString *> *labels = [matchingSnapshots valueForKeyPath:@"@distinctUnionOfObjects.label"];
   XCTAssertEqualObjects([NSSet setWithArray:labels], expectedLabels);
@@ -50,7 +55,7 @@
 
 - (void)testDescendantsMatchingXPath
 {
-  NSArray<XCElementSnapshot *> *matchingSnapshots = [self.testedView.lastSnapshot fb_descendantsMatchingXPathQuery:@"//XCUIElementTypeButton[@label='Alerts']"];
+  NSArray<XCElementSnapshot *> *matchingSnapshots = [self.testedView.fb_lastSnapshot fb_descendantsMatchingXPathQuery:@"//XCUIElementTypeButton[@label='Alerts']"];
   XCTAssertEqual(matchingSnapshots.count, 1);
   XCTAssertEqual(matchingSnapshots.lastObject.elementType, XCUIElementTypeButton);
   XCTAssertEqualObjects(matchingSnapshots.lastObject.label, @"Alerts");
@@ -61,46 +66,76 @@
   XCUIElement *button = self.testedApplication.buttons[@"Alerts"];
   XCTAssertTrue(button.exists);
   [button resolve];
-  XCElementSnapshot *windowSnapshot = [button.lastSnapshot fb_parentMatchingType:XCUIElementTypeWindow];
+  XCElementSnapshot *windowSnapshot = [button.fb_lastSnapshot fb_parentMatchingType:XCUIElementTypeWindow];
   XCTAssertNotNil(windowSnapshot);
   XCTAssertEqual(windowSnapshot.elementType, XCUIElementTypeWindow);
 }
 
+@end
+
+@interface XCElementSnapshotHelperTests_AttributePage : FBIntegrationTestCase
+@end
+
+@implementation XCElementSnapshotHelperTests_AttributePage
+
+- (void)setUp
+{
+  [super setUp];
+  static dispatch_once_t onceToken;
+  dispatch_once(&onceToken, ^{
+    [self launchApplication];
+    [self goToAttributesPage];
+  });
+}
+
 - (void)testParentMatchingOneOfTypes
 {
-  [self goToAttributesPage];
   XCUIElement *todayPickerWheel = self.testedApplication.pickerWheels[@"Today"];
   XCTAssertTrue(todayPickerWheel.exists);
   [todayPickerWheel resolve];
-  XCElementSnapshot *datePicker = [todayPickerWheel.lastSnapshot fb_parentMatchingOneOfTypes:@[@(XCUIElementTypeDatePicker), @(XCUIElementTypeWindow)]];
+  XCElementSnapshot *datePicker = [todayPickerWheel.fb_lastSnapshot fb_parentMatchingOneOfTypes:@[@(XCUIElementTypeDatePicker), @(XCUIElementTypeWindow)]];
   XCTAssertNotNil(datePicker);
   XCTAssertEqual(datePicker.elementType, XCUIElementTypeDatePicker);
 }
 
 - (void)testParentMatchingOneOfTypesWithXCUIElementTypeAny
 {
-  [self goToAttributesPage];
   XCUIElement *todayPickerWheel = self.testedApplication.pickerWheels[@"Today"];
   XCTAssertTrue(todayPickerWheel.exists);
   [todayPickerWheel resolve];
-  XCElementSnapshot *otherSnapshot = [todayPickerWheel.lastSnapshot fb_parentMatchingOneOfTypes:@[@(XCUIElementTypeAny), @(XCUIElementTypeWindow)]];
+  XCElementSnapshot *otherSnapshot = [todayPickerWheel.fb_lastSnapshot fb_parentMatchingOneOfTypes:@[@(XCUIElementTypeAny), @(XCUIElementTypeWindow)]];
   XCTAssertNotNil(otherSnapshot);
   XCTAssertEqual(otherSnapshot.elementType, XCUIElementTypeOther);
 }
 
 - (void)testParentMatchingOneOfTypesWithAbsentParents
 {
-  [self goToAttributesPage];
   XCUIElement *todayPickerWheel = self.testedApplication.pickerWheels[@"Today"];
   XCTAssertTrue(todayPickerWheel.exists);
   [todayPickerWheel resolve];
-  XCElementSnapshot *otherSnapshot = [todayPickerWheel.lastSnapshot fb_parentMatchingOneOfTypes:@[@(XCUIElementTypeTab), @(XCUIElementTypeLink)]];
+  XCElementSnapshot *otherSnapshot = [todayPickerWheel.fb_lastSnapshot fb_parentMatchingOneOfTypes:@[@(XCUIElementTypeTab), @(XCUIElementTypeLink)]];
   XCTAssertNil(otherSnapshot);
+}
+
+@end
+
+@interface XCElementSnapshotHelperTests_ScrollView : FBIntegrationTestCase
+@end
+
+@implementation XCElementSnapshotHelperTests_ScrollView
+
+- (void)setUp
+{
+  [super setUp];
+  static dispatch_once_t onceToken;
+  dispatch_once(&onceToken, ^{
+    [self launchApplication];
+    [self goToScrollPageWithCells:false];
+  });
 }
 
 - (void)testParentMatchingOneOfTypesWithFilter
 {
-  [self goToScrollPageWithCells:false];
   XCUIElement *threeStaticText = self.testedApplication.staticTexts[@"3"];
   [threeStaticText resolve];
   NSArray *acceptedParents = @[
@@ -108,16 +143,15 @@
                                @(XCUIElementTypeCollectionView),
                                @(XCUIElementTypeTable),
                                ];
-  XCElementSnapshot *scrollView = [threeStaticText.lastSnapshot fb_parentMatchingOneOfTypes:acceptedParents
-    filter:^(XCElementSnapshot *snapshot) {
-        return [snapshot isWDVisible];
-     }];
+  XCElementSnapshot *scrollView = [threeStaticText.fb_lastSnapshot fb_parentMatchingOneOfTypes:acceptedParents
+                                                                                        filter:^(XCElementSnapshot *snapshot) {
+                                                                                          return [snapshot isWDVisible];
+                                                                                        }];
   XCTAssertEqualObjects(scrollView.identifier, @"scrollView");
 }
 
 - (void)testParentMatchingOneOfTypesWithFilterRetruningNo
 {
-  [self goToScrollPageWithCells:false];
   XCUIElement *threeStaticText = self.testedApplication.staticTexts[@"3"];
   [threeStaticText resolve];
   NSArray *acceptedParents = @[
@@ -125,32 +159,47 @@
                                @(XCUIElementTypeCollectionView),
                                @(XCUIElementTypeTable),
                                ];
-  XCElementSnapshot *scrollView = [threeStaticText.lastSnapshot fb_parentMatchingOneOfTypes:acceptedParents
-    filter:^(XCElementSnapshot *snapshot) {
-        return NO;
-    }];
+  XCElementSnapshot *scrollView = [threeStaticText.fb_lastSnapshot fb_parentMatchingOneOfTypes:acceptedParents
+                                                                                        filter:^(XCElementSnapshot *snapshot) {
+                                                                                          return NO;
+                                                                                        }];
   XCTAssertNil(scrollView);
 }
 
 - (void)testDescendantsCellSnapshots
 {
-  [self goToScrollPageWithCells:false];
   XCUIElement *scrollView = self.testedApplication.scrollViews[@"scrollView"];
   [scrollView resolve];
   FBAssertWaitTillBecomesTrue(self.testedApplication.staticTexts[@"3"].fb_isVisible);
-  NSArray *cells = [scrollView.lastSnapshot fb_descendantsCellSnapshots];
+  NSArray *cells = [scrollView.fb_lastSnapshot fb_descendantsCellSnapshots];
   XCTAssertGreaterThanOrEqual(cells.count, 10);
   XCElementSnapshot *element = cells.firstObject;
   XCTAssertEqualObjects(element.label, @"0");
 }
 
+@end
+
+@interface XCElementSnapshotHelperTests_ScrollViewCells : FBIntegrationTestCase
+@end
+
+@implementation XCElementSnapshotHelperTests_ScrollViewCells
+
+- (void)setUp
+{
+  [super setUp];
+  static dispatch_once_t onceToken;
+  dispatch_once(&onceToken, ^{
+    [self launchApplication];
+    [self goToScrollPageWithCells:true];
+  });
+}
+
 - (void)testParentCellSnapshot
 {
-  [self goToScrollPageWithCells:true];
   FBAssertWaitTillBecomesTrue(self.testedApplication.staticTexts[@"3"].fb_isVisible);
   XCUIElement *threeStaticText = self.testedApplication.staticTexts[@"3"];
   [threeStaticText resolve];
-  XCElementSnapshot *xcuiElementCell = [threeStaticText.lastSnapshot fb_parentCellSnapshot];
+  XCElementSnapshot *xcuiElementCell = [threeStaticText.fb_lastSnapshot fb_parentCellSnapshot];
   XCTAssertEqual(xcuiElementCell.elementType, 75);
 }
 
